Recurrent payment based checks in upgrades refactored
The upgraders now check not only if the payment was paid via gateway
supporting recurrent payment, but also whether this recurrent payment
is still actually active or not.

The issue with this was that people with expiring/expired recurrent
payments weren't able to upgrade at all. We needed a way to guide them
to non-recurrent upgrade branches.