Swapping stopping of current and creation of new subscription when splitting
In original order, the old subscription was first stopped and
SubscriptionEndsEvent was emitted. Event handlers handling this couldn't
know that there would be another subscription created in a second
and acted as this was the actual end of running subscription.

This caused bonus subscription being generated occasionally during upgrade.

As the new subscription is now created first, the event handlers have
enough information to not act as if the subscription was actually ending.
They can check that there's already another subscription following.