Available upgraders now check all active subscriptions, not just one
The original implementaion checked only one subscription. After the
free promo upgrade it used to return this subscription without payment
which caused upgrade to halt.
All subscriptions are being checked now and first to match the rules is
used to upgrade.

Commit also fixes subscription splitting as the events weren't emitted
correctly and the promo subscription was not stopped when the "base"
subscription was stopped during splitting process.

remp/crm#877