Bugfixing subscription shortened handler using incorrect start_time
The start time of following subscription should be end time of previous
subscription. Previous implementation incorrectly assumed, that
subscription from event can be used, but in case of an upgrade this
was a subscription that was stopped. We therefore ignored new upgraded
subscription that was created.

Correct end time that should be used as a start time of subsequent
subscription is the one belonging to the newly created upgraded
subscription.

remp/crm#851