Fix the order of how subscription upgrade is done - first create a new subscription and then expire the former
This is due to Bonus2DaysSubscriptionHandler, which detects wheter to add 2 days bonus subscription based on if there are any actual ongoing subscriptions.
Therefore we need to upgrade by creating a new subscription and then shortening the previous to avoid adding 2 day bonus (which is not required in this case).

remp/crm#1429