Get rid of transactions during uprade process, causing issues
The transactions were incompatible with hermes events which
got handled sooner than the transaction was commited. As a results
the events were either thrown away or handled much later than
intended (thanks to the RetryTrait).

If we find a solution compatible with Hermes messages, we'll
incorporate them back.

remp/crm#1883