Remove subscription type limitation for subsequent upgrades
Subsequent upgrades are now able to resolve and upgrade ANY following
subscription, even if it is one of different type or length.

Commit also fixes an issue with subscription days calculation,
which could occur during subsequent upgrades of subscriptions
that are passing through the daylight savings time. Due to the
hour shift the original implementation would use a day less/more
for the calculation. New implementation derives the length
from the number of seconds.

remp/crm#2887