Fixed TrialUpgrade not resolving default subscription
The error appeared if omit_content config was expected to be used.
It was stored in the trial upgrade subscription, but it wasn't
passed to the resolved within the finalize() method. Resolver
wasn't able to get the target subscription type because of
not omitting content access to be omitted.

remp/crm#3489