Fix prices in Free/Paid recurrent upgraders
If trial subscription type with multiple periods were used,
upgrader incorrectly displayed higher price even when the
trial would still continue even after upgrade.

As for the FreeRecurrentUpgrade, there was an issue that the
custom amount of the full subscription type would be set
after upgrade to the trial subscription type even when the
user would still be eligible for the trial price. Since
RecurrentPaymentResolver is now capable of resolving the
correct price, there shouldn't be a need to use RP'scustom amount
at this place and is now being removed.

remp/helpdesk#3817